[C]: change subscription fragment_limit to size_t and make publication buffers const.

+++ aeron-client/src/main/c/aeron_subscription.c
@@ -315,14 +315,14 @@ int64_t aeron_subscription_channel_status(aeron_subscription_t *subscription)
 }
 
 int aeron_subscription_poll(
-    aeron_subscription_t *subscription, aeron_fragment_handler_t handler, void *clientd, int fragment_limit)
+    aeron_subscription_t *subscription, aeron_fragment_handler_t handler, void *clientd, size_t fragment_limit)
 {
     volatile aeron_image_list_t *image_list;
 
     AERON_GET_VOLATILE(image_list, subscription->conductor_fields.image_lists_head.next_list);
 
     size_t length = image_list->length;
-    int fragments_read = 0;
+    size_t fragments_read = 0;
     size_t starting_index = subscription->round_robin_index++;
     if (starting_index >= length)
     {
@@ -331,28 +331,30 @@ int aeron_subscription_poll(
 
     for (size_t i = starting_index; i < length && fragments_read < fragment_limit; i++)
     {
-        fragments_read += aeron_image_poll(image_list->array[i], handler, clientd, fragment_limit - fragments_read);
+        fragments_read += (size_t)aeron_image_poll(
+            image_list->array[i], handler, clientd, fragment_limit - fragments_read);
     }
 
     for (size_t i = 0; i < starting_index && fragments_read < fragment_limit; i++)
     {
-        fragments_read += aeron_image_poll(image_list->array[i], handler, clientd, fragment_limit - fragments_read);
+        fragments_read += (size_t)aeron_image_poll(
+            image_list->array[i], handler, clientd, fragment_limit - fragments_read);
     }
 
     aeron_subscription_propose_last_image_change_number(subscription, image_list->change_number);
 
-    return fragments_read;
+    return (int)fragments_read;
 }
 
 int aeron_subscription_controlled_poll(
-    aeron_subscription_t *subscription, aeron_controlled_fragment_handler_t handler, void *clientd, int fragment_limit)
+    aeron_subscription_t *subscription, aeron_controlled_fragment_handler_t handler, void *clientd, size_t fragment_limit)
 {
     volatile aeron_image_list_t *image_list;
 
     AERON_GET_VOLATILE(image_list, subscription->conductor_fields.image_lists_head.next_list);
 
     size_t length = image_list->length;
-    int fragments_read = 0;
+    size_t fragments_read = 0;
     size_t starting_index = subscription->round_robin_index++;
     if (starting_index >= length)
     {
@@ -361,19 +363,19 @@ int aeron_subscription_controlled_poll(
 
     for (size_t i = starting_index; i < length && fragments_read < fragment_limit; i++)
     {
-        fragments_read += aeron_image_controlled_poll(
+        fragments_read += (size_t)aeron_image_controlled_poll(
             image_list->array[i], handler, clientd, fragment_limit - fragments_read);
     }
 
     for (size_t i = 0; i < starting_index && fragments_read < fragment_limit; i++)
     {
-        fragments_read += aeron_image_controlled_poll(
+        fragments_read += (size_t)aeron_image_controlled_poll(
             image_list->array[i], handler, clientd, fragment_limit - fragments_read);
     }
 
     aeron_subscription_propose_last_image_change_number(subscription, image_list->change_number);
 
-    return fragments_read;
+    return (int)fragments_read;
 }
 
 long aeron_subscription_block_poll(
